Added functions to game_object and fixed memory leaks
Added set_global_rotation, get_global_rotaion and set_global_orbit to game_object to make it easier to set variables that are not effected by the perents of the object. With the memory leaks I found out that I wasn't managing them at all and they were staying there.