Created colliders for all planets and deleted player class.
The previous error was to do with the collisers being set to nullptr and therefor the position being set to 0,0. This was the same spawn position as the player and caused it to stop being valid as a result. The removing of the player class was to do with that it was inherating and had a new class for one extra variable that could be in the game_object class anyway.