FIxed collisions and memory leaks.
Problems with collisions was to do with the radus being set to the with of the game_object and not half it so the colider was much bigger then it was meant to be. The problems with the memory leaks was to do with not deleting the colliders or the player.